Adjust random variable bias strength (#2633)
* fix issue #2480

* internally make the random variable bias strength 10X bigger

* internally make the random variable bias strength 1000X bigger

* use std::nth_element() instead of std::partial_sort() to save some computation cycles.

* trigger clang-format hook

* trigger clang-format

* trigger clang-format

* clang-format pre-push hook

* Tweaked bias strength multiplier to give "better" results.

* address Scott concern

* trigger clang-format

* trigger clang-format

* trigger clang-format

* trigger clang-format

+++ include/vapor/FlowParams.h
@@ -105,8 +105,8 @@ public:
     std::string GetRakeBiasVariable() const;
     void        SetRakeBiasVariable(const std::string &);
 
-    float GetRakeBiasStrength() const;
-    void  SetRakeBiasStrength(float);
+    long GetRakeBiasStrength() const;
+    void SetRakeBiasStrength(long);
 
     int  GetPastNumOfTimeSteps() const;
     void SetPastNumOfTimeSteps(int);
